+++ src/spillway/estimators/__init__.py
@@ -1,0 +1,1 @@
+"""Predicting what a request will cost, before it is made."""